requirements.txt
python-telegram-bot==20.7
flask==3.0.0
requests==2.31.0
pymongo==4.6.1
python-dotenv==1.0.0
APScheduler==3.10.4
cryptography==41.0.8
